+++ html_vault/__init__.py
@@ -1,3 +1,3 @@
 """HTML Vault package."""
 
-__version__ = "0.4.10"
+__version__ = "0.4.11"